+++ castep_outputs/utilities/utility.py
@@ -3,33 +3,14 @@ Utility functions for parsing castep outputs
 """
 
 import collections.abc
-import fileinput
-import functools
-import json
-import logging
-import pprint
+from collections import defaultdict
 import re
 from collections import defaultdict
 from copy import copy
 from typing import (Any, Callable, Dict, List, Mapping, MutableMapping, Optional, TextIO, Tuple,
                     TypeVar, Union, Type, Iterable)
 
-import castep_outputs.castep_res as REs
-
-from .filewrapper import FileWrapper
-
-T = TypeVar("T")
-
-_YAML_TYPE: Optional[str]
-try:
-    from ruamel import yaml
-    _YAML_TYPE = "ruamel"
-except ImportError:
-    try:
-        import yaml
-        _YAML_TYPE = "yaml"
-    except ImportError:
-        _YAML_TYPE = None
+import castep_outputs.utilities.castep_res as REs
 
 
 def normalise_string(string: str) -> str:
@@ -125,56 +106,7 @@ def flatten_dict(dictionary: MutableMapping[Any, Any],
     return dict(items)
 
 
-def json_dumper(data: Any, file: TextIO):
-    """ Basic JSON format dumper """
-    json.dump(data, file, indent=2)
-
-
-def ruamel_dumper(data: Any, file: TextIO):
-    """ Basic ruamel.yaml format dumper """
-    yaml_eng = yaml.YAML(typ='unsafe')
-    yaml_eng.dump(data, file)
-
-
-def yaml_dumper(data: Any, file: TextIO):
-    """ Basic yaml format dumper """
-    yaml.dump(data, file)
-
-
-def pprint_dumper(data: Any, file: TextIO):
-    """ PPrint dumper """
-    print(pprint.pformat(data), file=file)
-
-
-def print_dumper(data: Any, file: TextIO):
-    """ Print dumper """
-    print(data, file=file)
-
-
-SUPPORTED_FORMATS = {"json": json_dumper,
-                     "ruamel": ruamel_dumper,
-                     "yaml": yaml_dumper,
-                     "pprint": pprint_dumper,
-                     "print": print_dumper}
-
-
-def get_dumpers(dump_fmt: str) -> Callable:
-    """
-    Get appropriate dump for unified interface
-    """
-    if dump_fmt not in SUPPORTED_FORMATS:
-        raise ValueError(f"Cannot output in {dump_fmt} format.")
-
-    if dump_fmt == "yaml":
-        if _YAML_TYPE is None:
-            raise ImportError("Couldn't find valid yaml dumper (ruamel.yaml / yaml)"
-                              "please install and try again.")
-        dump_fmt = _YAML_TYPE
-
-    return SUPPORTED_FORMATS[dump_fmt]
-
-
-def stack_dict(out_dict: MutableMapping[Any, Any], in_dict: Mapping[Any, Any]):
+def stack_dict(out_dict: Dict[Any, List], in_dict: Dict[Any, List]) -> Dict[Any, List]:
     """ Append items in `in_dict` to the keys in `out_dict` """
     for key, val in in_dict.items():
         out_dict[key].append(val)
